Extract `Presenter` type
Extracts all presentation logic for `Vulnerability` and `Warning`s into
its own self-contained type.
